+++ src/lang-xslt/main/uk/co/reecedunn/intellij/plugin/xslt/intellij/javaee/XsltResourceProvider.kt
@@ -16,16 +16,22 @@
 package uk.co.reecedunn.intellij.plugin.xslt.intellij.javaee
 
 import com.intellij.javaee.ResourceRegistrar
+import com.intellij.javaee.ResourceRegistrarImpl
 import com.intellij.javaee.StandardResourceProvider
 import uk.co.reecedunn.intellij.plugin.xslt.intellij.lang.XSLT
 
 class XsltResourceProvider : StandardResourceProvider {
     companion object {
+        private const val XSLT_20_URI = "https://www.w3.org/2007/schema-for-xslt20.xsd"
         private const val XSLT_30_URI = "https://www.w3.org/TR/xslt-30/schema-for-xslt30.xsd"
     }
 
     override fun registerResources(registrar: ResourceRegistrar?) {
-        registrar!!.addStdResource(XSLT_30_URI, "/schemas/xslt-3_0.xsd", XsltResourceProvider::class.java)
-        registrar.addStdResource(XSLT.NAMESPACE, "3.0", "/schemas/xslt-3_0.xsd", XsltResourceProvider::class.java)
+        val resources = registrar as ResourceRegistrarImpl
+
+        resources.addInternalResource(XSLT_20_URI, "xslt-2_0.xsd")
+        resources.addStdResource(XSLT_30_URI, "/schemas/xslt-3_0.xsd", XsltResourceProvider::class.java)
+
+        resources.addStdResource(XSLT.NAMESPACE, "3.0", "/schemas/xslt-3_0.xsd", XsltResourceProvider::class.java)
     }
 }